workaround: index.js: Use wrong actions path to receive OpenVR input.
Even though the controllers are rendered and emit a valid pose,
it is required to set an incorrect bindings path to make A-Painter
receive any actions.
Funnily an incorrect path must be set, an empty one will result in a JS
backtrace and the scene not loading at all.
SteamVR will complain about a missing bindings file and suggest to
create one in the VR UI, but it works fine without setting anything.

The bahaviour was broken in `25d1d6d14a84e63b525bedd0c2693df77abb2627`,
where the bindings path was fixed.

The core problem is that apparently the actions don't provide what A-Painter
expects, didn't test any other apps.

Same behaviour for Vive Wand and Valve Index controllers.